+++ rentplace/src/main/java/kattsyn/dev/rentplace/services/impl/FavouritesServiceImpl.java
@@ -1,7 +1,7 @@
 package kattsyn.dev.rentplace.services.impl;
 
 import jakarta.transaction.Transactional;
-import kattsyn.dev.rentplace.dtos.PropertyDTO;
+import kattsyn.dev.rentplace.dtos.properties.PropertyDTO;
 import kattsyn.dev.rentplace.entities.Property;
 import kattsyn.dev.rentplace.entities.User;
 import kattsyn.dev.rentplace.exceptions.NotFoundException;
